move files to the right place
